feat(authorization): 新增账户管理模块及相关枚举常量
- 新增账户相关枚举类：AccountEnum、AccountPermissionEnum、AccountStatusEnum
- 新增账户角色关联DTO及Repository
- 新增账户用户关联DTO及Repository
- 修改LoginAccount实体类账户状态字段类型为Integer
- 新增LoginAccountApi控制器，提供账户增删改查接口
- 新增LoginAccountService业务类，实现账户管理逻辑
- 修改LoginAccountRepository，添加账户状态批量更新等方法
- 调整包结构，将登录相关API和服务移至login子包
- 更新验证码异常消息键名
- 设置本地化语言为zh_CN

+++ model/src/main/java/com/ranyk/model/business/login/entity/LoginAccount.java
@@ -48,7 +48,7 @@ public class LoginAccount implements Serializable {
      * 登录账户状态: -2: 停用/注销; -1: 禁用; 0: 锁定; 1: 启用;
      */
     @Column(name = "login_account_status", nullable = false, columnDefinition = "TINYINT DEFAULT 0 COMMENT '登录账户状态: -2: 停用/注销; -1: 禁用; 0: 锁定; 1: 启用;'")
-    private byte accountStatus;
+    private Integer accountStatus;
     /**
      * 数据创建时间
      */
